requirements.txt
google-adk>=1.0.0
pytest>=7.0.0
requests>=2.28.0